Handle errors from recv() - see Bug #1527930
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-3-patches@15411 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/smux/smux.c
@@ -1264,7 +1264,7 @@ smux_snmp_process(int exact,
                   size_t * return_len, u_char * return_type, int sd)
 {
     u_char          packet[SMUXMAXPKTSIZE], *ptr, result[SMUXMAXPKTSIZE];
-    size_t          length = SMUXMAXPKTSIZE;
+    int             length = SMUXMAXPKTSIZE;
     u_char          type;
     size_t          packet_len;
